Adding folder for the first Question.
The function dataTypes() will return diferent values depending on the
data type of the parameter. The test cases developed for it are also
included in this commit. For example, if the input is a string, the
function will return the length of the string and if its a number, it
will tell you its comparism with respect to 100 etc.