This commit is for the second question
Create a function to print all the prime numbers from 0 to the input
value. The functions will eliminate numbers less than 2 from being
excecution. It will also iterate only over the odd numbers. This will
reduce the time it takes to go through the list by half.

The tests folder also includes 11 test cases that checks for the
correctness of the function.